sample_agent/flight_booking_gemini/requirements.txt
google-genai
python-dotenv
termcolor
